Вынес константы в отдельный модуль
