+++ src/const.js
@@ -3,3 +3,18 @@ export const sortType = {
   DATE: 'date',
   RATING: 'rating',
 };
+
+export const FilmControlAction = {
+  watchlist: 'toggleWatchList',
+  watched: 'toggleWatched',
+  favorite: 'toggleFavotite',
+};
+
+export const FILMS_PER_ROW = 5;
+
+export const RankTrie = {
+  0: '',
+  10: 'Novice',
+  20: 'Fan',
+  21: 'Movie buff',
+};
